general: - unify/consolidate afk tracking logic and fix bugs related to it vvv-serverside porting: - server functions: cvars, svhelp, masterz, fps, dump, afklist, poll, cp, swapteams, lockname, lockteam, mute, amstatus, amkick - d/bs analysis & logging - client commands: send tstats if teamoverlay == 2 - returns are now in impressive count client: - teamoverlay 2. get tstats data. do basic parsing (but no display yet)

+++ code/game/g_defrag.h
@@ -129,10 +129,10 @@ typedef struct savedPosition_s {
 		// pers.
 		struct {
 			struct {
-				float		lasthurtcarrier;
-				float		lastreturnedflag;
-				float		flagsince;
-				float		lastfraggedcarrier;
+				int			lasthurtcarrier;
+				int			lastreturnedflag;
+				int			flagsince;
+				int			lastfraggedcarrier;
 			} teamState;
 			struct {
 				//int	startLevelTime; // this stuff is more for demo cutting so no need to compensate.
